requirements.txt
# Core Django and dependencies
Django==5.2.0
django-advanced-password-validation
django-cache-cleaner==0.2.0
django-import-export==4.3.7
django-unfold==0.60.0
psycopg2-binary>=2.9.3

# Environment management
python-dotenv

# Data analysis and notebooks
ipykernel
jupyter
notebook
pandas

# Code quality and pre-commit hooks
isort==6.0.1
pre-commit==4.2.0
